+++ tilequeue/constants.py
@@ -1,0 +1,1 @@
+MAX_TILE_ZOOM = 16  # tilezen's default max supported zoom